♻️ rename AIcon
