+++ src/components/UI/ExternalLink.vue
@@ -1,20 +1,12 @@
 <template>
   <a :class="$style.container">
-    <icon name="mdi:open-in-new" :class="$style.icon" />
+    <a-icon name="mdi:open-in-new" :class="$style.icon" />
     <slot />
   </a>
 </template>
 
-<script lang="ts">
-import { defineComponent } from 'vue'
-import Icon from '/@/components/UI/Icon.vue'
-
-export default defineComponent({
-  name: 'LinkWithIcon',
-  components: {
-    Icon
-  }
-})
+<script lang="ts" setup>
+import AIcon from './AIcon.vue'
 </script>
 
 <style lang="scss" module>
